refactor: 쿠폰 모델 util 파일 분리
